requirements.txt
pandas==2.1.4
sweetviz==2.3.1
streamlit==1.36.0
streamlit-option-menu==0.3.13
python-dotenv==1.0.1
pymongo==4.8.0
pycaret==3.3.2
pygwalker==0.4.9